Make FriendsList a "popup"; GameSearch now respects user query when showing results * For now pulling data from dummy data Updated app styling to account for new features
