+++ frontend/src/App.css
@@ -95,6 +95,41 @@ select {
   color: var(--color-dark-blue);
 }
 
+.modal {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.6);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+}
+
+.modalContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: var(--color-white);
+  padding: 1em;
+  border-radius: 5px;
+  box-shadow: 0 0 5px var(--color-black);
+  z-index: 1001;
+  width: 50%;
+  height: 50%;
+}
+
+.modalContainer .modalHeader {
+  display: flex;
+  align-self: flex-start;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  margin-bottom: 1em;
+}
+
 /* For buttons */
 button {
   border: none;
@@ -237,7 +272,6 @@ select:disabled {
   /* We could use transitions to only show this error styling for a little bit to alert
     the user then hide it to enable normal input accessibility */
 }
-
 /* End of text-like <input> elements */
 
 main {
@@ -418,6 +452,9 @@ form {
   font-family: "Inter", sans-serif;
 }
 
+.hidden {
+  display: none;
+}
 /* End of utility classes */
 
 /* For the <img> elements */
@@ -545,6 +582,35 @@ form {
 }
 
 /*Friends List*/
+#openFriendsListBtn {
+  position: fixed;
+  bottom: 0;
+  right: 0;
+  padding: 1rem;
+  background-color: var(--color-gray);
+  border: 1px solid var(--color-white);
+  border-radius: 10px;
+  box-shadow: 0 0 10px 2px var(--color-dark-blue);
+  z-index: 100;
+}
+/* Make it into a popup on the bottom right */
+.friends-list {
+  position: fixed;
+  bottom: 0;
+  right: 0;
+  width: 40vh;
+  height: 50vh;
+  padding: 1rem;
+  background-color: var(--color-black);
+  border: 1px solid var(--color-white);
+  border-radius: 10px;
+  box-shadow: 0 0 10px 5px var(--color-gray);
+  opacity: 1;
+  z-index: 100;
+  overflow-y: scroll;
+  overflow-x: hidden;
+}
+
 .friends-list-body .flexDirectionColumn {
   margin-bottom: 10px;
   /* padding: 5px; */
