+++ frontend/src/routes/GameSearch.jsx
@@ -12,6 +12,7 @@ export default function GameSearch(props) {
     });
 
     const getSearchResults = (e) => {
+        alert("Searching for games with name " + searchState.query);
         e.preventDefault();
         // TODO: Make this actually search the database
         fetch("/dummyData.json", {
@@ -28,19 +29,16 @@ export default function GameSearch(props) {
         }).then((res) => {
             if (res.status === 200) {
                 res.json().then((data) => {
-                    let searchResults = [];
+                    // let searchResults = [];
                     // TODO: Filter the results based on the filter rules
                     // Select only the results that match the query
                     // Right now, it just filters based on the player's username
                     // If the query is empty, return all results
-                    if (searchState.query !== "" && data?.users?.length > 0) {
-                        searchResults = data.users?.filter((player) => {
-                            return (
-                                player.username.toLowerCase().includes(searchState.query.toLowerCase())
-                            )
+                    if (searchState.query !== "" && data.games) {
+                        const res = data.games.filter((game) => {
+                            return game.name.toLowerCase().includes(searchState.query.toLowerCase());
                         });
-
-                        setSearchState((prev) => ({ ...prev, results: searchResults }));
+                        setSearchState((prev) => ({ ...prev, results: res }));
                     }
                 });
             } else {
@@ -57,7 +55,7 @@ export default function GameSearch(props) {
             <hr className="width-100" />
             <div className="flexDirectionRow">
                 {/* Search Bar */}
-                <LabeledInput type="text" id="search" label="" defaultValue={searchState.search} containerStyle={{ flexGrow: 1 }} orientation="horizontal" onChange={(e) => setSearchState((prev) => ({ ...prev, search: e.target.value }))} />
+                <LabeledInput type="text" id="search" label="" placeholder="Search for a game..." defaultValue={searchState.search} containerStyle={{ flexGrow: 1 }} orientation="horizontal" onChange={(e) => setSearchState((prev) => ({ ...prev, query: e.target.value }))} />
                 {/* Search Button */}
                 <button onClick={(e) => getSearchResults(e)}>Search</button>
             </div>
@@ -65,11 +63,15 @@ export default function GameSearch(props) {
                 {/* Main Games Container */}
                 <div className="flexDirectionRow justifyContentSpaceBetween" style={{ width: "75%", margin: "0 auto" }}>
                     {/* TODO: Fix this so the container doesn't render if not needed */}
-                    {searchState.results?.games?.length > 0 && (<><div>Search Results:</div>
-                    <div>{searchState.results?.games.length || 0} games found.</div></>)}
+                    {searchState.results?.length > 0 && (
+                        <>
+                            <div>Search Results:</div>
+                            <div>{searchState.results?.length || 0} games found.</div>
+                        </>
+                    )}
                 </div>
                 <div className="flexDirectionColumn flexWrap">
-                    {searchState.results?.games?.map((game) => (
+                    {searchState.results?.map((game) => (
                         // TODO: Extract styling for GameCard and allow it to vary for GameSearch and when it's used in profiles
                         <GameCard key={game.GameId} game={game} withPlayers={true} />
                     )) || <div>No games found.</div>}
